+++ cmake/Config.cmake
@@ -36,6 +36,12 @@ if (MSVC)
     # https://msdn.microsoft.com/en-us/library/ms175759.aspx
     add_definitions (-D_CRT_SECURE_CPP_OVERLOAD_STANDARD_NAMES=1 -D_CRT_SECURE_CPP_OVERLOAD_STANDARD_NAMES_COUNT=1)
 
+    # Disable code analysis warnings about constant constant comparisons. A
+    # lot of our template functions trigger this when they do things like:
+    #
+    # if (T == traits<U>::some_value) { ... } else { ... }
+    add_compile_options(/wd6326)
+
     # Enable standards-conformance mode for MSVC compilers that support this
     # flag (Visual C++ 2017 and later).
     #
@@ -152,6 +158,7 @@ cxx_add_compile_options(Clang
     --std=c++11
     -Wall
     -Werror
+    -Wno-null-dereference
     -Wno-unknown-warning-option
     -Wno-unused-local-typedefs)
 
@@ -161,6 +168,7 @@ cxx_add_compile_options(AppleClang
     --std=c++11
     -Wall
     -Werror
+    -Wno-null-dereference
     -Wno-unknown-warning-option
     -Wno-unused-local-typedefs)