+++ cpp/test/comm/idl.cpp
@@ -1,3 +1,0 @@
-#include "comm_test_common_apply.cpp"
-#include "comm_test_common_comm.cpp"
-#include "comm_test_common_types.cpp"