+++ cs/src/core/protocols/SimpleBinary.cs
@@ -29,8 +29,8 @@
                      string, wstring        | count | characters |
                                             '-------'------------'
 
-                           count            variable encoded uint32 count of 1-byte (for
-                                            string) or 2-byte (for wstring) Unicode code
+                           count            uint32 count of 1-byte (for string)
+                                            or 2-byte (for wstring) Unicode code
                                             units (variable encoded in v2)
 
                            characters       1-byte UTF-8 code units (for string) or 2-byte
@@ -52,6 +52,13 @@
                             count           uint32 count of {key,mapped} pairs (variable encoded in v2)
 
                             key, mapped     each item encoded according to its type
+
+                                           .-------. .-----------.
+                    bonded                 | count | | marshaled |
+                                           '-------' '-----------'
+                            count           uint32 count of bytes (always fixed-width, even in v2)
+
+                            marshaled       a marshaled payload
 
 */
 #endregion